#78: Move the corresponding interfaces (like Arg) to their packages
