+++ src/main/java/io/github/dgroup/term4j/arg/ArgOf.java
@@ -23,7 +23,6 @@
  */
 package io.github.dgroup.term4j.arg;
 
-import io.github.dgroup.term4j.Arg;
 import org.cactoos.Scalar;
 import org.cactoos.scalar.True;
 import org.cactoos.scalar.Unchecked;
