+++ src/main/java/io/github/dgroup/term4j/arg/FileOf.java
@@ -23,7 +23,6 @@
  */
 package io.github.dgroup.term4j.arg;
 
-import io.github.dgroup.term4j.Arg;
 import java.io.File;
 import java.util.List;
 import org.cactoos.Scalar;
