+++ src/main/java/io/github/dgroup/term4j/arg/LongOf.java
@@ -24,7 +24,6 @@
 
 package io.github.dgroup.term4j.arg;
 
-import io.github.dgroup.term4j.Arg;
 import java.util.List;
 import org.cactoos.Func;
