+++ src/main/java/io/github/dgroup/term4j/arg/PathOf.java
@@ -23,7 +23,6 @@
  */
 package io.github.dgroup.term4j.arg;
 
-import io.github.dgroup.term4j.Arg;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
