+++ src/main/java/io/github/dgroup/term4j/arg/TextOf.java
@@ -23,7 +23,6 @@
  */
 package io.github.dgroup.term4j.arg;
 
-import io.github.dgroup.term4j.Arg;
 import java.util.List;
 import org.cactoos.Func;
 import org.cactoos.Text;
